Remove duplicated code and uncessary cleanup
