+++ packages/rum-core/src/domain/resource/resourceCollection.spec.ts
@@ -48,7 +48,6 @@ describe('resourceCollection', () => {
 
     registerCleanupTask(() => {
       startResult.stop()
-      rawRumEvents = []
     })
   }
