+++ packages/rum-core/src/domain/view/viewCollection.spec.ts
@@ -113,7 +113,6 @@ describe('viewCollection', () => {
 
     registerCleanupTask(() => {
       collectionResult.stop()
-      rawRumEvents = []
     })
   }
