+++ packages/rum-core/src/domain/vital/vitalCollection.spec.ts
@@ -25,7 +25,6 @@ describe('vitalCollection', () => {
     rawRumEvents = collectAndValidateRawRumEvents(lifeCycle)
 
     registerCleanupTask(() => {
-      rawRumEvents = []
       clock.cleanup()
     })
   })
