+++ packages/rum-core/test/index.ts
@@ -7,5 +7,4 @@ export * from './mockRumSessionManager'
 export * from './noopRecorderApi'
 export * from './testSetupBuilder'
 export * from './emulate/mockPerformanceObserver'
-export * from './eventFormatValidation'
 export * from './mockPageStateHistory'
